Move Dimensions test to right file, there not in view model anymore
